Apply suggestions from code review

+++ front/rule.common.php
@@ -83,8 +83,7 @@ if (isset($_POST["action"])) {
     $start = microtime(true);
 
    // Limit computed from current time
-    /** @var number $max */
-    $max = get_cfg_var("max_execution_time");
+    $max = (int) get_cfg_var("max_execution_time");
     $max = $start + ($max > 0 ? $max / 2.0 : 30.0);
 
     Html::header(
